backend/pyproject.toml
[tool.poetry]
name = "safetasks-backend"
version = "0.1.0"
description = "SafeTasks V2 Backend API"
authors = ["SafeTasks Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.104.1"
uvicorn = {extras = ["standard"], version = "^0.40.0"}
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.1"
pydantic-settings = "^2.4.0"
asyncpg = "^0.31.0"
python-dotenv = "^1.2.1"
greenlet = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
email-validator = "^2.3.0"
bcrypt = "4.0.1"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
python-multipart = "^0.0.21"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry-core/build-backend"
